WIP for #10: Use settings.py and config.js
This is a WIP. It's not finished. It's mainly an idea of how to use
settings.py and honcho .env file for app settings.

It's working in standalone mode, i.e, for users. CI hasn't been
completed. However, user initialization needs to be fixed.

When CI customization is needed, env variables should be written to
{{repo}}/.env from tpl/init/sd-x.sh. Settings.py should do the rest.

config.js still needs to be improved/thought.