fireq - porting to modern [lxc] :hammer: :rocket:
Finally we are pleased to announce the ***Next Gen of fireq CI system V2*** which comes alot of performance improvements and other a few features including publishing an instance and stateful snapshots. Special thanks to all of the systeam, particularly Fawad and Abbas for taking this challenging projet, working out the optimal configuration for compilation, debugging and thorough testing.